Move generic mail class to other file
