+++ src/controllers/generic-mail-controller.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from "express";
 import { GenericMail } from "../types/types";
-import MailService from "../services/mail-service";
+import GenericMailService from "../services/generic-mail-service";
 
-export const sendMail = async (req: Request, res: Response) => {
+export const sendGenericMail = async (req: Request, res: Response) => {
   try {
     const messageOptions : GenericMail = {
       to: req.body.to,
@@ -12,7 +12,7 @@ export const sendMail = async (req: Request, res: Response) => {
       images: req.body.images,
     };
 
-    const response = await new MailService().sendGenericMail(messageOptions);
+    const response = await new GenericMailService().sendGenericMail(messageOptions);
     return res.status(200).json(response);
   } catch (error) {
     res.status(400).json({ error });
