+++ src/routes/base-router.ts
@@ -1,5 +1,5 @@
 import { Express } from "express";
-import mail from "./mail-router";
+import mail from "./generic-mail-routes";
 
 const setBaseRouter = (app: Express): void => {
   app.use("/mail", [mail]);
