+++ src/services/mail-service.ts
@@ -1,8 +1,7 @@
 import Mail from "nodemailer/lib/mailer";
 import logger from "../utils/logger";
 import transporter from "../config/mailer";
-import { Base64Image, GenericMail } from "types";
-import { getCurrentHourMinute } from "../utils/time";
+import { Base64Image } from "../types/types";
 
 class MailService {
   async base64ImagesToAttachment(images: Array<Base64Image>) {
@@ -22,34 +21,6 @@ class MailService {
     });
 
     return attachments;
-  }
-
-  async sendGenericMail(messageOptions: GenericMail) {
-    const formattedTime = getCurrentHourMinute();
-
-    let attachments: Array<Mail.Attachment> = [];
-    if (messageOptions.images) {
-      attachments = await this.base64ImagesToAttachment(messageOptions.images);
-    }
-
-    const mailOptions = {
-      to: messageOptions.to,
-      from: process.env["SMTP_FROM"],
-      subject: messageOptions.subject,
-      template: "index",
-      context: {
-        preheader: messageOptions.message.substring(0, 100),
-        partialName: "generic-message",
-        content: {
-          user: messageOptions.user,
-          time: formattedTime,
-          message: messageOptions.message,
-        },
-      },
-      attachments: attachments,
-    };
-
-    await this.sendMail(mailOptions);
   }
 
   async sendMail(mailOptions: Mail.Options) {
